refactor: make mixin instead
